change datype for lat lng price to float.

+++ backend/db/migrations/2-create-spot.js
@@ -40,11 +40,11 @@ module.exports = {
         allowNull: false,
       },
       lat: {
-        type: Sequelize.DECIMAL(20,10),
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       lng: {
-        type: Sequelize.DECIMAL(20,10),
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       name: {
@@ -57,7 +57,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.DECIMAL(20,10),
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       createdAt: {
